Customizing Clerk implementation, removing cotter
