+++ components/authed-only.js
@@ -1,21 +1,20 @@
 import {useEffect, useState} from "react";
-import { SignedIn, SignedOut, RedirectToSignIn } from '@clerk/nextjs';
+import { SignedIn, SignedOut, SignIn, useUser } from '@clerk/nextjs';
+import SignoutLink from "./navs/signout-link";
+import SigninForm from "./navs/signin-form";
+import LogUser from "./log-user";
 
 export default function AuthedOnly({children}) {
-    // Display the login page
-    return (
-        <main>
-            <SignedIn>
-              {children}
-            </SignedIn>
-            <SignedOut>
-              <RedirectToSignIn />
-                <div style={{textAlign: 'center'}}>
-                    <h1>Writer Portal</h1>
-                    <p>This portal is exclusively for <a href="https://draft.dev/">Draft.dev</a> writers.</p>
-                    <p>Log in using the same email you used to apply.</p>
-                </div>
-            </SignedOut>
-        </main>
-    );
+  return (
+    <main>
+      <SignedIn>
+        <LogUser />
+        {children}
+        <p><SignoutLink /></p>
+      </SignedIn>
+      <SignedOut>
+        <SigninForm />
+      </SignedOut>
+    </main>
+  );
 }
