+++ components/layout.js
@@ -1,7 +1,6 @@
 // components/Layout.js
 import React, { Component } from "react";
 import Head from "next/head";
-import { UserButton } from "@clerk/clerk-react";
 
 export default class Layout extends Component {
   render() {
@@ -25,7 +24,7 @@ export default class Layout extends Component {
             </a>
           </div>
           <div className="nav-right">
-            <UserButton />
+            <a href="/"></a>
           </div>
         </header>
         {children}
