+++ package.json
@@ -11,11 +11,9 @@
     "node": ">=12.22"
   },
   "dependencies": {
+    "@clerk/clerk-react": "^2.6.0",
     "@clerk/nextjs": "^2.6.0",
     "airtable": "^0.11.0",
-    "cotter": "^0.3.33",
-    "cotter-node": "^0.1.3",
-    "cotter-token-js": "^0.1.11",
     "dayjs": "^1.10.4",
     "lodash": "^4.17.21",
     "logrocket": "^2.1.2",
