+++ pages/api/assignments/[assignmentId]/index.js
@@ -1,25 +1,13 @@
-import {CotterAccessToken} from "cotter-token-js";
 const {Pool} = require("pg");
 const connectionString = process.env.PG_CONNECTION_STRING;
 const pool = new Pool({connectionString});
+import { requireSession, users } from "@clerk/nextjs/api";
 
-export default async (req, res) => {
-    const {assignmentId} = req.query;
-
-    // Check that the authorization header exists
-    if (!("authorization" in req.headers)) {
-        res.statusCode = 401;
-        return res.end("Authorization header missing");
-    }
-
-    // Extract the token string
-    const auth = await req.headers.authorization;
-    const bearer = auth.split(" ");
-    const token = bearer[1];
-
+export default requireSession(async (req, res) => {
     try {
-        // Decode the Cotter JWT, "decoded.payload.identifier" is the user's email
-        const decoded = new CotterAccessToken(token);
+        const {assignmentId} = req.query;
+        // Get user from Clerk API
+        const user = await users.getUser(req.session.userId);
 
         if (req.method === "GET") {
             // Get assignment by assignmentId
@@ -44,7 +32,7 @@ export default async (req, res) => {
                                where $2 = ANY (writer_email)
                            ) as your_requests on your_requests.id = ANY (assignments.requests)
                            where assignments.id like $1;`;
-            const {rows} = await pool.query(query, [assignmentId, decoded.payload.identifier]);
+            const {rows} = await pool.query(query, [assignmentId, user.emailAddresses[0].emailAddress]);
 
             // Respond with results
             res.statusCode = 200;
@@ -56,4 +44,4 @@ export default async (req, res) => {
         res.statusCode = 500;
         res.end("Server error. Something went wrong.");
     }
-};
+});
