+++ pages/api/assignments/available.js
@@ -1,23 +1,12 @@
-import {CotterAccessToken} from "cotter-token-js";
 const { Pool } = require("pg");
 const connectionString = process.env.PG_CONNECTION_STRING;
 const pool = new Pool({ connectionString });
+import { requireSession, users } from "@clerk/nextjs/api";
 
-export default async (req, res) => {
-  // Check that the authorization header exists
-  if (!("authorization" in req.headers)) {
-    res.statusCode = 401;
-    return res.end("Authorization header missing");
-  }
-
-  // Extract the token string
-  const auth = await req.headers.authorization;
-  const bearer = auth.split(" ");
-  const token = bearer[1];
-
+export default requireSession(async (req, res) => {
   try {
-    // Decode the Cotter JWT, "decoded.payload.identifier" is the user's email
-    const decoded = new CotterAccessToken(token);
+    // Get user from Clerk API
+    const user = await users.getUser(req.session.userId);
 
     if (req.method === "GET") {
       // Get assignments
@@ -43,7 +32,7 @@ export default async (req, res) => {
                      and assignments.writer = '{}'
                      group by assignments.id, your_requests.request_date
                      order by assignments.writer_due_date asc;`;
-      const { rows } = await pool.query(query, ['Assigning', decoded.payload.identifier]);
+      const { rows } = await pool.query(query, ['Assigning', user.emailAddresses[0].emailAddress]);
 
       // Respond with results
       res.statusCode = 200;
@@ -55,4 +44,4 @@ export default async (req, res) => {
     res.statusCode = 500;
     return res.end("Server error. Something went wrong.");
   }
-};
+});
